+++ src/components/Passports/IconSpotlight.svelte
